pyproject.toml
[project]
name = "discordbotcito"
version = "0.1.0"
description = "Discord music bot with autoplay and Opus streaming"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord.py[voice]>=2.6.0,<3.0.0",
    "yt-dlp>=2025.1.0",
    "ytmusicapi>=1.11.0,<2.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "agno>=2.3.26",
    "mcp",
    "sqlalchemy",
    "openai>=2.15.0",
    "textual>=7.3.0",
]

[project.scripts]
audit = "audit:run_audit"

[tool.uv]
package = true

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["audit*", "game_agent*"]
